requirements.txt
build~=1.0.3
channels[daphne]~=4.0.0
coverage~=7.4.0
Django~=4.2.9
djangorestframework~=3.14.0
djangorestframework-stubs[compatible-mypy]~=3.14.5
django-stubs[compatible-mypy]~=4.2.7
flake8~=6.1.0
mypy~=1.7.1
pokerkit~=0.4.10
Sphinx~=7.2.6
twine~=4.0.2
